requirements.txt
comtypes==1.1.11
pypiwin32==223
python-decouple==3.6
pyttsx3==2.90
pywin32==304
